Adds shell script for quickly adding ready-to-edit person file (see description for how-to).
To use, first pull the repo and have the script file handy. Double clicking on the file should bring up a command prompt screen, asking for some input. Here, you can put in the person ID of the person for whom you want a file created and press Enter.
The shell script then does several things: it (1) repeats back the person ID you've just entered (2) checks if the person ID you've entered is valid (whether it follows the 3 letters 2 numbers scheme) (3) checks if there's an existing file for the entered person ID (4) If there is, then it automatically fixes the current person ID to not conflict and says what that new ID is (5) If everything goes well, you should have a ready-to-edit file in the people directory. If things did not go well, there'll be an error message for you to read. If the script freezes at some point, try running it through the terminal, see below.
From here, you can go to your oXygen editor, navigate to the Open/Find tab, type in the person ID (the new person ID if the one you entered conflicted with an existing file), and open the file. You can of course, open the new file from File Explorer as well.

Another way to use it is directly through the Git Bash terminal. This tends to be more reliable and quicker for repeated use. First, you navigate to the folder, then right click a space not occupied by a file to bring up a menu. In this menu, find the "Open Git Bash here" option and click on that. From there, type "./quick-new.sh" and press Enter. The rest of the process is the same as if you started it by double clicking.

The neat thing about this script is that you only need to worry about the 3 letters. The number part is automatically handled for you since the script knows what numbers already exist. So for example, you could type in SMI01 and it will automatically detect and create SMI17 for you.

Feel free to email me if you have any questions regarding this.