Updates quick-new.sh for extra functionalities (see description)
The script now understands 3 letters input e.g. if you enter only "SMI" in the input, it will automatically work out a person code based on that. Compared to before, this saves the two keystrokes needed for the number part of the person ID.
After the new file is created, the script will attempt to open said new file with Oxygen, saving the time and effort spent on manually searching and opening that file. The opening itself might take a while. That's just Oxygen donig its thing.
Finally, the script is a tiny bit more readable for those interesting in the nuts and bolts :)

As always, feel free to email me for any questions you may have.