Makes script fully POSIX compliant and adds feature to automatically fill encoder initials, based on Git user.name value.
It should be a tiny bit faster now that it's POSIX compliant and can be run with speed-optimised shells like dash, and it's more portable as a matter of course. If you don't know what POSIX is, it's not that important from a user perspective so it's nothing to worry about.

The new auto-fill feature for encoder initials only works if you've set a Git user.name value. Make sure to capitalise properly and have a space between your first and last name. Instructions for how to set up a Git user.name can be found in the tutorial session manual.