Turns DAR02.xml into placeholder, moves DAR02 content to DAR04, changes year file to correspond (for details see description).
A DAR02.html file exists and is referenced in the year file. However, this a blank page and there is no accompanying XML file (henceforth, this ghost XML is referred to as old DAR02). This being the case, when a new "DAR" file was created, it got the number 02. This is, of course, a problem because now we have year files which point to the wrong person. I found this out just now so I'm fixing it as best I can, by shifting the new DAR file to an unoccupied number and editing the year file to reflect this.

We might need to do an automated scan to check this is the case for any more files. I haven't pushed it yet but this is the second time I've encountered the, for the lack of a better term, "ghost XML phenomenon". The first time is with DAV04, which will be committed soon.